+++ src/hrflow_connectors/connectors/hrflow/warehouse/__init__.py
@@ -1,0 +1,4 @@
+from hrflow_connectors.connectors.hrflow.warehouse.job import HrFlowJobWarehouse  # noqa
+from hrflow_connectors.connectors.hrflow.warehouse.profile import (  # noqa
+    HrFlowProfileWarehouse,
+)
