+++ src/hrflow_connectors/connectors/hrflow/warehouse/job.py
@@ -6,8 +6,8 @@ from logging import LoggerAdapter
 from hrflow import Hrflow
 from pydantic import BaseModel, Field
 
-from hrflow_connectors.connectors.hrflow.schemas import HrFlowJob, HrFlowProfile
-from hrflow_connectors.core import Warehouse, WarehouseReadAction, WarehouseWriteAction
+from hrflow_connectors.connectors.hrflow.schemas import HrFlowJob
+from hrflow_connectors.core import Warehouse, WarehouseWriteAction
 
 LIST_JOBS_LIMIT = 30
 
@@ -42,17 +42,6 @@ class WriteJobParameters(BaseModel):
     enrich_with_parsing: bool = Field(
         False, description="When enabled jobs are enriched with HrFlow.ai parsing"
     )
-
-
-class ReadProfileParameters(BaseModel):
-    api_secret: str = Field(
-        ...,
-        description="X-API-KEY used to access HrFlow.ai API",
-        repr=False,
-    )
-    api_user: str = Field(..., description="X-USER-EMAIL used to access HrFlow.ai API")
-    source_key: str = Field(..., description="HrFlow.ai source key")
-    profile_key: str = Field(..., description="HrFlow.ai profile key")
 
 
 def remove_html_tags(text: str) -> str:
@@ -275,38 +264,8 @@ def write(
     return failed_jobs
 
 
-def read(adapter: LoggerAdapter, parameters: ReadProfileParameters) -> t.List[t.Dict]:
-    hrflow_client = Hrflow(
-        api_secret=parameters.api_secret, api_user=parameters.api_user
-    )
-    response = hrflow_client.profile.indexing.get(
-        source_key=parameters.source_key, key=parameters.profile_key
-    )
-    if "Unable to find object" in response["message"]:
-        adapter.info(
-            "No profile found for source_key={} profile_key={} response={}".format(
-                parameters.source_key, parameters.profile_key, response
-            )
-        )
-        return []
-    elif response["code"] >= 400:
-        adapter.error(
-            "Failed to get profile source_key={} profile_key={} response={}".format(
-                parameters.source_key, parameters.profile_key, response
-            )
-        )
-        raise Exception("Failed to get profile")
-    return [response["data"]]
-
-
 HrFlowJobWarehouse = Warehouse(
     name="HrFlow.ai Jobs",
     data_schema=HrFlowJob,
     write=WarehouseWriteAction(parameters=WriteJobParameters, function=write),
-)
-
-HrFlowProfileWarehouse = Warehouse(
-    name="HrFlow.ai Profiles",
-    data_schema=HrFlowProfile,
-    read=WarehouseReadAction(parameters=ReadProfileParameters, function=read),
 )
